:sparkles: uses classes as modules and export them on index
