+++ SpriteGL/index.ts
@@ -1,0 +1,10 @@
+import Shader from "./Shader";
+import SpriteRenderer from "./SpriteRenderer";
+import TextDrawer from "./TextDrawer";
+import VBO from "./VBO";
+
+const SpriteGL = { Shader, SpriteRenderer, TextDrawer, VBO };
+
+export { Shader, SpriteGL, SpriteRenderer, TextDrawer, VBO };
+
+export default SpriteGL;
